requirements/common.txt
aiofiles==0.7.0
aiohttp
argcomplete==1.11.0
backoff==1.11.1
boto3==1.36.2
cachetools==5.2.0
dacite==1.6.0
Deprecated==1.2.11
dill==0.3.6
ftfy==6.1.1
httpx==0.23.0
hypercorn==0.13.2
kaleido==0.2.1
matplotlib==3.5.2
mongoengine~=0.29.1
motor~=3.6.0
packaging==20.3
pandas>=1.3
plotly==5.17.0
pprintpp==0.4.0
psutil>=5.7.0
pymongo~=4.9.2
pytz==2022.1
PyYAML==6.0.1
regex==2022.8.17
retrying>=1.3.3
scikit-learn>=0.23.2
scikit-image>=0.16.2
setuptools>=45.2.0,<71
sseclient-py>=1.7.2
sse-starlette>=0.10.3
starlette==0.40.0
strawberry-graphql==0.257.0
tabulate==0.8.10
wcmatch==8.3
websocket-client==1.8.0
xmltodict==0.12.0
yarl
